requirements.txt
certifi==2024.8.30
cffi==1.16.0
cycler==0.12.1
decorator==5.1.1
imageio==2.35.1
importlib-resources==6.4.0
kiwisolver==1.4.5
matplotlib==3.8.4
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
Pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyWavelets==1.6.0
scikit-image==0.23.2
scipy==1.13.1
tifffile==2024.8.30
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.5
zipp==3.19.2